refactor : infrastructure 중간 추상화와 JPA 구현체 패키지 분리
