+++ oop-infrastructure/src/main/kotlin/com/jaeyong/oop/infrastructure/health/jpa/HealthJpaRepository.kt
@@ -1,6 +1,7 @@
-package com.jaeyong.oop.infrastructure.health.repository
+package com.jaeyong.oop.infrastructure.health.jpa
 
 import com.jaeyong.oop.infrastructure.health.entity.HealthEntity
+import com.jaeyong.oop.infrastructure.health.repository.HealthEntityRepository
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface HealthJpaRepository : JpaRepository<HealthEntity, Long>, HealthEntityRepository
