+++ oop-infrastructure/src/main/kotlin/com/jaeyong/oop/infrastructure/user/jpa/UserJpaRepository.kt
@@ -1,6 +1,7 @@
-package com.jaeyong.oop.infrastructure.user.repository
+package com.jaeyong.oop.infrastructure.user.jpa
 
 import com.jaeyong.oop.infrastructure.user.entity.UserEntity
+import com.jaeyong.oop.infrastructure.user.repository.UserEntityRepository
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface UserJpaRepository : JpaRepository<UserEntity, Long>, UserEntityRepository {
